Fix client routes and restore portal functionality
